+++ AssignmentFiles/myRectangle.h
@@ -11,7 +11,7 @@ public:
 	myRectangle();
 	myRectangle(myPoint UL, myPoint UR, myPoint LR, myPoint LL);
 	myRectangle(const myPoint& upperLeft, const myPoint& lowerRight);
-	void scale(int factor);
+	void scale(double factor);
 	//void rotate(double angle, int units);
 	//void translate(const myVector& u);
 	void draw(CDC* pDC);
